requirements.txt
aiohttp==3.10.8
pytest-asyncio==0.24.0
